+++ R/read_write_plum.R
@@ -131,7 +131,7 @@ read.dets.plum <- function(core, coredir, n.supp=c(), date.sample, sep=",", dec=
           if(!is.na(dets[2,n]))
             if(dets[2,n] != "")
               nsupp.infile <- dets[2,n]
-        if(length(dets[3,n]) > 0) # 3rd, which radium case to use to estimate supported
+        if(length(dets[3,n]) > 0) # calib.plot3rd, which radium case to use to estimate supported
           if(!is.na(dets[3,n]))
             if(dets[3,n] != "")
               racase.infile <- dets[3,n]
@@ -403,7 +403,7 @@ Plum.cleanup <- function(set=get('info')) {
 
 
 # read in default values, values from previous run, any specified values, and report the desired one. Internal function. Differs from Bacon.settings because it requires Pb-specific settings.
-.plum.settings <- function(core, coredir, dets, thick, remember=TRUE, d.min, d.max, d.by, depths.file,
+.plum.settings <- function(core, coredir, dets, detsBacon, thick, remember=TRUE, d.min, d.max, d.by, depths.file,
   slump, acc.mean, acc.shape, mem.mean, mem.strength, boundary, hiatus.depths, hiatus.max, hiatus.shape,
   BCAD, cc, postbomb, cc1, cc2, cc3, cc4, depth.unit, normal, t.a, t.b, delta.R, delta.STD, prob,
   defaults, runname, ssize, dark, MinAge, MaxAge, cutoff, age.res, after, age.unit,
@@ -450,9 +450,12 @@ Plum.cleanup <- function(set=get('info')) {
 
   if(is.na(d.min) || d.min=="NA")
     d.min <- min(dets[,4])
-  if(is.na(d.max) || d.max=="NA")
-    #d.max <- max(dets[,4])
-    d.max <- max(dets[,4]) # tmp
+  if(is.na(d.max) || d.max=="NA") {
+    #d.max <- max(dets[,4]) # tmp
+    if(length(detsBacon) == 0)  # Dec 2022. Should this only be if ra.case=0?
+      d.max <- max(dets[1:(nrow(dets)-n.supp),4]) else
+        d.max <- max(dets[1:(nrow(dets)-n.supp),4], detsBacon[,4])
+    }
   if(length(acc.shape) < length(acc.mean))
     acc.shape <- rep(acc.shape, length(acc.mean)) else
       if(length(acc.shape) > length(acc.mean))
@@ -578,6 +581,13 @@ write.plum.file <- function(set=get('info')) {
     dets[1,1] <- NA # calling this "d.min" causes issues
     dets[1,3] <- max(1e5, 1e3*dets[,4], 1e3*dets[,3])
     dets[1,depthColumn] <- set$d.min
+    bg <- set$supportedData[,3] # now reassign depths of the data in the tail
+    # tail measurements do not require depths, so we reassign them
+    if(set$ra.case < 2) # or <2?
+      for(i in 1:length(bg)) {
+          d_i <- which(dets[,depthColumn] == bg[i])
+          dets[d_i,depthColumn] <- set$d.min+1
+        }
   }
 
   if(is.na(set$d.max) || set$d.max > max(dets[,depthColumn])) { # repeat relevant row, change error and depth